Changes to PatchModels, and add some optional functions (edge_difference, blenging) and their sample programs.

+++ modules/OpenCV/include/BackgroundSubtractAlgorithm.h
@@ -4,15 +4,15 @@
 #include <cv.h>
 
 namespace skl{
-	class BackgroundSubtractAlgorithm:public FilterMat2Mat<int>{
+	class BackgroundSubtractAlgorithm:public FilterMat2Mat<double>{
 		public:
 			BackgroundSubtractAlgorithm(){};
 			virtual ~BackgroundSubtractAlgorithm(){};
-			virtual int compute(const cv::Mat& src, cv::Mat& dest){
+			virtual double compute(const cv::Mat& src, cv::Mat& dest){
 				cv::Mat mask = cv::Mat(src.size(),CV_8UC1,255);
 				return compute(src, mask, dest);
 			}
-			virtual int compute(const cv::Mat& src,const cv::Mat& mask, cv::Mat& dest)=0;
+			virtual double compute(const cv::Mat& src,const cv::Mat& mask, cv::Mat& dest)=0;
 			virtual void updateBackgroundModel(const cv::Mat& bg,const cv::Mat& mask)=0;
 	};
